pyproject.toml
[project]
name = "maf-2-the-max"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-azure-ai>=1.0.0b251114",
    "agent-framework-core>=1.0.0b251114",
    "aiohttp>=3.13.2",
    "packaging>=25.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "fastmcp>=2.13.0.2",
    "agent-framework-devui>=1.0.0b251120",
]
